+++ NEWS.md
@@ -1,5 +1,6 @@
 1.3.6 (DD/MM/YY)
 -----
+* Do not confuse warnings with errors on shapefile output (#87)
 
 1.3.5 (19/02/13)
 -----